re #21 making sure the buffer windows are as large as the query.  To avoid recreating buffers on every iteration, we use buffers that have been created on the previous step if they are the same size.
At first I was going to take out this buffer functionality entirely except I remembered that by creating the buffers, we get to say what the output data type is that's going into the vectorize dataset function.  It's otherwise whatever type readasarray declares and it is not always consistent between calls.